Convert landing.go into api
